+++ rom.ino
@@ -2,11 +2,11 @@
 
 void initROM() {
   EEPROM.begin(EEPROM_SIZE);
-  int countReboot = EEPROM.read(0); // Get current reboot count
+  uint32_t countReboot = EEPROM.read(0); // Get current reboot count
   EEPROM.write(0, countReboot + 1); // Update reboot count
   EEPROM.commit();                  // Save to ROM
 }
 
-int getROMData(int index) {
+uint32_t getROMData(size_t index) {
   return EEPROM.read(index);
 }
